split up http adapter tests
The main http.spec.js was almost 1000 lines long and by splitting these
tests up it makes maintenance a little easier. In the process of
splitting up also moved to do things in a more ES6-y way.
